Fix analog input for toggles, fix buttontype inputMax
Analog input was causing out of range values for settings such as Monster Counter and Button Prompt Style.
This results in [!] being displayed, and the setting itself becoming invalid if no digital input is available to correct it. 
This commit also fixes the Button Type being stuck on `XBOX`, by removing the subtraction of the os_type from the inputMax value.